[metadata] add support for monospace property
