+++ go.mod
@@ -3,6 +3,6 @@ module github.com/go-text/typesetting
 go 1.17
 
 require (
-	github.com/go-text/typesetting-utils v0.0.0-20230326210548-458646692de6
-	golang.org/x/image v0.3.0
+	github.com/go-text/typesetting-utils v0.0.0-20230411134717-a95abaf3f778
+	golang.org/x/image v0.7.0
 )
