+++ opentype/api/metadata/descriptor_test.go
@@ -35,8 +35,23 @@ func TestMetadata(t *testing.T) {
 		ld, err := loader.NewLoader(bytes.NewReader(f))
 		tu.AssertNoErr(t, err)
 
-		gotAspect, gotFamily := Metadata(ld)
-		tu.AssertC(t, gotAspect == test.aspect, fmt.Sprint(gotAspect))
-		tu.AssertC(t, gotFamily == test.family, gotFamily)
+		got := Metadata(ld)
+		tu.AssertC(t, got.Aspect == test.aspect, fmt.Sprint(got.Aspect))
+		tu.AssertC(t, got.Family == test.family, got.Family)
 	}
+}
+
+func Test_isMonospace(t *testing.T) {
+	for _, file := range tu.Filenames(t, "common") {
+		f, err := td.Files.ReadFile(file)
+		tu.AssertNoErr(t, err)
+
+		ld, err := loader.NewLoader(bytes.NewReader(f))
+		tu.AssertNoErr(t, err)
+
+		fd := newFontDescriptor(ld)
+		tu.Assert(t, td.Monospace[file] == fd.isMonospace())
+	}
+
+	tu.Assert(t, !(&fontDescriptor{}).isMonospace()) // check it does not crash
 }
